Pool: Refactor pool extension to its own file.
